+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+Development:
+  - wait until any attestations for the current slot have completed before shutting down
+
 1.4.0:
   - increase accuracy of beacon block proposal scorer by incorporating attestation history
   - multinode submitter returns after first successful submission rather than waiting for all to complete